Finally got my butt in gear to fix the intial estimation.
I actaully feel pretty good about this, it uses a resize senseing function
for the contraint plot to adjust linewidth sizese to the size of the
figure, additionally adjusting the location of the legend based on size fo
the figure to fit in the graph.

Two graphs are output by power_constr.m:
	the first, by cnstr.m, is the contraint graph with 8 seleactable
	constraints.
	the second, by power_constr.m creates a minimum power requried for
	various manuevers.
Notes: Need to fix the power requried, as I think I am outputting Power
used, i.e., I am putting the manuever power, not the power avalibe
neccesary for the power required... if that makes sense.